# Fix perf_analyzer report for ensemble models