Fixes #2429 .
Perf analyzer will report the queue and compute time as the sum of its composing models. This PR makes report more comprehensible and less confusing.

### Before
```
  Server: 
    Inference count: 677
    Execution count: 677
    Successful request count: 677
    Avg request latency: 8348 usec
    Total avg compute input time : 82 usec
    Total avg compute infer time : 7875 usec
    Total avg compute output time : 223 usec
    Total avg queue time : 0 usec

  Composing models: 
  dali, version: 
      Inference count: 677
      Execution count: 677
      Successful request count: 677
      Avg request latency: 2307 usec (overhead 3 usec + queue 42 usec + compute input 5 usec + compute infer 2220 usec + compute output 37 usec)

  resnet50_fp32_libtorch, version: 
      Inference count: 677
      Execution count: 677
      Successful request count: 677
      Avg request latency: 6000 usec (overhead 48 usec + queue 36 usec + compute input 76 usec + compute infer 5654 usec + compute output 186 usec)
```

### After
```
  Server: 
    Inference count: 678
    Execution count: 678
    Successful request count: 678
    Avg request latency: 8312 usec (overhead 98 usec + queue 90 usec + compute 8124 usec)

  Composing models: 
  dali, version: 
      Inference count: 678
      Execution count: 678
      Successful request count: 678
      Avg request latency: 2223 usec (overhead 3 usec + queue 56 usec + compute input 6 usec + compute infer 2123 usec + compute output 35 usec)

  resnet50_fp32_libtorch, version: 
      Inference count: 678
      Execution count: 678
      Successful request count: 678
      Avg request latency: 6044 usec (overhead 50 usec + queue 34 usec + compute input 75 usec + compute infer 5728 usec + compute output 157 usec)
```